bitcoin_stream/requirements.txt
Flask==1.0.2
kafka-python==2.0.0
redis==3.4.1
